services/ai-broker/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
httpx>=0.27.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.25.0
python-multipart>=0.0.6
